event: move event/ package into internal/
Move the event package from top-level event/ directory into internal/event/ to follow Go convention for internal packages.

- relocated event/ files to internal/event/
- updated imports in 8 files across proxy/, internal/logmon/, and root
