+++ internal/event/event.go
